ex01 lightly refactored

+++ ex01/iter.hpp
@@ -1,19 +1,15 @@
 #pragma once
 
-#include <cstddef>
-
 template <typename T>
-
-void	iter(T *arr, const size_t length, void (*f)(T&))
+void	iter(T *arr, const unsigned int length, void (*f)(T&))
 {
-	for (size_t i = 0; i < length; i++)
+	for (unsigned int i = 0; i < length; i++)
 		f(arr[i]);
 }
 
 template <typename T>
-
-void	iter(T *arr, const size_t length, void (*f)(const T&))
+void	iter(T *arr, const unsigned int length, void (*f)(const T&))
 {
-	for (size_t i = 0; i < length; i++)
+	for (unsigned int i = 0; i < length; i++)
 		f(arr[i]);
 }
